+++ CHANGELOG.md
@@ -74,9 +74,11 @@ got ctrl z to function and changed keybinds
 - brightness adjustments now work as intended
 - more bugs
 - fixed dock canvas placement bug
+# 2.2.8
+- video support (mp4 only currently)
+- uses nearest neighbor for upscaling
 # future updates will be slow and may not happen at all, but could include:
 - longer multiframe support
-- mp4 support
 - canvasAPI for modders if eli doesnt just make one himself lol (sorry)
 - separate mod for art acessibility (allows artists to use mypaint to edit in-game canvases or new ones)
 - canvas locking for hosts